refactor(shared): move meta into shared
