+++ src/app/core/footer/footer-copyright/footer-copyright.component.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component, input } from '@angular/core';
 
-import { MetaPage } from '../../models/meta.model';
+import { MetaPage } from '@awg-shared/meta/meta.model';
 
 /**
  * The FooterCopyright component.
