+++ src/app/core/footer/footer.component.ts
@@ -2,9 +2,8 @@ import { Component } from '@angular/core';
 
 import { LogoComponent } from '@awg-shared/logos/logo.component';
 import { LOGOS_DATA } from '@awg-shared/logos/logos.data';
-
-import { META_DATA } from '../data/meta.data';
-import { MetaSectionTypes } from '../models/meta.model';
+import { META_DATA } from '@awg-shared/meta/meta.data';
+import { MetaSectionTypes } from '@awg-shared/meta/meta.model';
 
 import { FooterCopyrightComponent } from './footer-copyright/footer-copyright.component';
 import { FooterDeclarationComponent } from './footer-declaration/footer-declaration.component';
