+++ src/app/shared/meta/meta-identifier-badges/meta-identifier-badges.component.scss
