+++ src/app/views/contact-view/contact-view.component.ts
@@ -1,10 +1,10 @@
 import { DatePipe } from '@angular/common';
 import { ChangeDetectionStrategy, Component, signal } from '@angular/core';
 
-import { META_DATA } from '@awg-core/data/meta.data';
-import { MetaSectionTypes } from '@awg-core/models/meta.model';
 import { HeadingComponent } from '@awg-shared/heading/heading.component';
-import { MetaIdentifierBadgesComponent } from '@awg-shared/meta-identifier-badges/meta-identifier-badges.component';
+import { MetaIdentifierBadgesComponent } from '@awg-shared/meta/meta-identifier-badges/meta-identifier-badges.component';
+import { META_DATA } from '@awg-shared/meta/meta.data';
+import { MetaSectionTypes } from '@awg-shared/meta/meta.model';
 
 /**
  * The ContactView component.
