+++ src/app/views/home-view/home-view.component.ts
@@ -1,10 +1,10 @@
 import { ChangeDetectionStrategy, Component, signal } from '@angular/core';
 import { RouterLink } from '@angular/router';
 
-import { META_DATA } from '@awg-core/data/meta.data';
-import { MetaSectionTypes } from '@awg-core/models/meta.model';
 import { AlertInfoComponent } from '@awg-shared/alert-info/alert-info.component';
 import { HeadingComponent } from '@awg-shared/heading/heading.component';
+import { META_DATA } from '@awg-shared/meta/meta.data';
+import { MetaSectionTypes } from '@awg-shared/meta/meta.model';
 import { ACTIVE_EDITION_SECTION_IDS } from '@awg-views/edition-view/data/active-edition-sections.data';
 import { EDITION_ROUTE_CONSTANTS } from '@awg-views/edition-view/edition-route-constants';
 import { EditionSectionLink } from '@awg-views/edition-view/models';
